+++ components/Message/Link/index.tsx
@@ -1,23 +1,10 @@
 import React from 'react';
 import classNames from 'classnames';
 import styles from './index.module.css';
+import { isImage, isUrlValid } from './utilities';
 
 interface Props {
   value: string;
-}
-
-const SUPPORTED_EXTENSIONS = ['png', 'jpg', 'jpeg', 'gif'];
-
-function isImage(href: string): boolean {
-  const extension = href.split('.').pop();
-  if (!extension) {
-    return false;
-  }
-  return SUPPORTED_EXTENSIONS.includes(extension.toLowerCase());
-}
-
-function isUrlValid(url: string): boolean {
-  return !url.startsWith('http://-') && !url.startsWith('https://-');
 }
 
 export default function Link({ value }: Props) {
